requirements.txt
ansible==2.8.2
PyYAML==5.3.1
